Allow inta-package imports with dummy-dependencies
Does not work on Python 2.

Allows you to specify multiple file paths that will be converted
to modules + paths to import them from. Each module imported will
be able to see the others. Anything else is still mocked.

Due to the complexity of Python's import mechanism and major
changes between Python 2 and 3 I have not been able to get this working
across both.